requirements.txt
click==8.0.3
dataclasses==0.8
Flask==2.0.1
greenlet==1.1.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.19.5
pandas==1.1.5
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.27
typing-extensions==3.10.0.2
uWSGI==2.0.20
Werkzeug==2.0.2
XlsxWriter==3.0.2
zipp==3.6.0
